Gave SourceControlConfigurationService its own file
